Tests moved to test folder
